+++ src/tracking/index.js
@@ -2,5 +2,15 @@
 
 const UsageEvent = require('./usage-event');
 const UsageTracker = require('./usage-tracker');
+const { UsageStore, createStore } = require('./usage-store');
+const MemoryStore = require('./stores/memory-store');
+const FilesystemStore = require('./stores/filesystem-store');
 
-module.exports = { UsageEvent, UsageTracker };
+module.exports = {
+  UsageEvent,
+  UsageTracker,
+  UsageStore,
+  createStore,
+  MemoryStore,
+  FilesystemStore,
+};